+++ OnyxScoutApplication/Shared/Models/FluentValidations/FieldValidator.cs
@@ -16,21 +16,21 @@ namespace OnyxScoutApplication.Shared.Models.FluentValidations
             RuleFor(x => x.NumericDefaultValue).Must(BeNumberBetween)
                 .WithMessage((model, s) =>
                     "Default value must be a number between " + model.MinValue + " and " + model.MaxValue)
-                .When(c => c.FieldType == FieldType.Numeric);
+                .When(c => c.FieldType == FieldType.Integer);
             RuleFor(x => x.Options).Must(v => v.Count >= 2).WithMessage("At least two options are required")
                 .When(c => c.FieldType == FieldType.OptionSelect);
             RuleFor(x => x.Options).Must(v => !v.Any(string.IsNullOrEmpty))
                 .WithMessage("Make sure all options are not empty").When(c => c.FieldType == FieldType.OptionSelect);
         }
 
-        private static bool BeNumberBetween(FieldDto model, int? value)
+        private static bool BeNumberBetween(FieldDto model, float? value)
         {
             if (value == null)
             {
                 return true;
             }
 
-            int v = (int) value;
+            float v = (float) value;
             return v >= model.MinValue && v <= model.MaxValue;
         }
     }
